fail early, refactor parallel code

+++ parallel.go
@@ -42,6 +42,7 @@ func NewParallelSyncBucket(opts ParallelBucketOptions, b Bucket) Bucket {
 }
 
 func (b *parallelBucketImpl) Push(ctx context.Context, local, remote string) error {
+	ctx, cancel := context.WithCancel(ctx)
 	files, err := walkLocalTree(ctx, local)
 	if err != nil {
 		return errors.WithStack(err)
@@ -59,21 +60,27 @@ func (b *parallelBucketImpl) Push(ctx context.Context, local, remote string) err
 		go func() {
 			defer wg.Done()
 			for fn := range in {
+				select {
+				case <-ctx.Done():
+					return
+				default:
+				}
+
 				if b.dryRun {
 					continue
 				}
 
 				err = b.Bucket.Upload(ctx, filepath.Join(remote, fn), filepath.Join(local, fn))
 				if err != nil {
-					grip.Error(err)
 					catcher.Add(err)
+					cancel()
 				}
 			}
 		}()
 	}
 	wg.Wait()
 
-	if b.deleteOnSync && !b.dryRun {
+	if ctx.Err() == nil && b.deleteOnSync && !b.dryRun {
 		catcher.Add(errors.Wrapf(os.RemoveAll(local), "problem removing '%s' after push", local))
 	}
 
@@ -81,6 +88,7 @@ func (b *parallelBucketImpl) Push(ctx context.Context, local, remote string) err
 
 }
 func (b *parallelBucketImpl) Pull(ctx context.Context, local, remote string) error {
+	ctx, cancel := context.WithCancel(ctx)
 	iter, err := b.List(ctx, remote)
 	if err != nil {
 		return errors.WithStack(err)
@@ -88,22 +96,18 @@ func (b *parallelBucketImpl) Pull(ctx context.Context, local, remote string) err
 
 	catcher := grip.NewBasicCatcher()
 	items := make(chan BucketItem)
-	toDelete := make(chan string)
+	toDelete := []string{}
 
-	deleteSignal := make(chan struct{})
 	go func() {
 		defer close(items)
 
 		for iter.Next(ctx) {
 			if iter.Err() != nil {
-				err = errors.Wrap(iter.Err(), "problem iterating bucket")
-				grip.Error(err)
-				catcher.Add(err)
+				catcher.Add(errors.Wrap(iter.Err(), "problem iterating bucket"))
 				return
 			}
 			select {
 			case <-ctx.Done():
-				grip.Error(ctx.Err())
 				catcher.Add(ctx.Err())
 				return
 			case items <- iter.Item():
@@ -113,65 +117,43 @@ func (b *parallelBucketImpl) Pull(ctx context.Context, local, remote string) err
 	}()
 
 	wg := &sync.WaitGroup{}
-
 	for i := 0; i < b.size; i++ {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
 			for item := range items {
+				select {
+				case <-ctx.Done():
+					catcher.Add(ctx.Err())
+					return
+				default:
+				}
+
 				name, err := filepath.Rel(remote, item.Name())
 				if err != nil {
-					err = errors.Wrap(err, "problem getting relative filepath")
-					grip.Error(err)
-					catcher.Add(err)
+					catcher.Add(errors.Wrap(err, "problem getting relative filepath"))
+					cancel()
 					return
 				}
 				localName := filepath.Join(local, name)
 				if err := b.Download(ctx, item.Name(), localName); err != nil {
-					grip.Error(err)
 					catcher.Add(err)
+					cancel()
 					return
-				}
-				select {
-				case <-ctx.Done():
-					grip.Error(ctx.Err())
-					catcher.Add(ctx.Err())
-					return
-				case toDelete <- item.Name():
-					continue
 				}
 			}
 		}()
 	}
+	wg.Wait()
 
-	go func() {
-		wg.Wait()
-		close(toDelete)
-	}()
-
-	go func() {
-		defer close(deleteSignal)
-		keys := []string{}
-		for key := range toDelete {
-			keys = append(keys, key)
-		}
-
-		if b.deleteOnSync {
-			if b.dryRun {
-				grip.Debug(message.Fields{
-					"dry_run": true,
-					"keys":    keys,
-					"message": "would delete after push",
-				})
-			} else {
-				catcher.Add(errors.Wrapf(b.RemoveMany(ctx, keys...), "problem removing '%s' after pull", remote))
-			}
-		}
-	}()
-
-	select {
-	case <-ctx.Done():
-	case <-deleteSignal:
+	if b.deleteOnSync && b.dryRun {
+		grip.Debug(message.Fields{
+			"dry_run": true,
+			"keys":    toDelete,
+			"message": "would delete after push",
+		})
+	} else if b.deleteOnSync {
+		catcher.Add(errors.Wrapf(b.RemoveMany(ctx, toDelete...), "problem removing '%s' after pull", remote))
 	}
 
 	return catcher.Resolve()
